+++ src/Knp/Menu/MenuItem.php
@@ -774,49 +774,99 @@ class MenuItem implements ItemInterface
     }
 
     /**
-     * Renders an array of label => uri pairs ready to be used for breadcrumbs.
+     * Renders an array ready to be used for breadcrumbs.
+     *
+     * Each element in the array will be an array with 3 keys:
+     * - `label` containing the label of the item
+     * - `url` containing the url of the item (may be `null`)
+     * - `item` containing the original item (may be `null` for the extra items)
      *
      * The subItem can be one of the following forms
      *   * 'subItem'
-     *   * array('subItem' => '@homepage')
+     *   * Knp\Menu\ItemInterface object
+     *   * array('subItem' => '
+     * @homepage')
      *   * array('subItem1', 'subItem2')
+     *   * array(array('label' => 'subItem1', 'url' => '@homepage'), array('label' => 'subItem2'))
      *
-     * @example
-     * // drill down to the Documentation menu item, then add "Chapter 1" to the breadcrumb
-     * $arr = $menu['Documentation']->getBreadcrumbsArray('Chapter 1');
-     * foreach ($arr as $name => $url)
-     * {
-     *
-     * }
-     *
-     * @param  mixed $subItem A string or array to append onto the end of the array
+     * @param mixed $subItem A string or array to append onto the end of the array
+     * @param boolean $strict Internal flag to optimize the lookup in parent nodes
+     * @throws \InvalidArgumentException if an element of the subItem is invalid
      * @return array
      */
-    public function getBreadcrumbsArray($subItem = null)
+    public function getBreadcrumbsArray($subItem = null, $strict = false)
     {
         $breadcrumbs = array();
-        $obj = $this;
 
-        if ($subItem) {
-            if (!is_array($subItem)) {
-                $subItem = array((string) $subItem => null);
+        if ($strict) {
+            $breadcrumbs = $subItem;
+        } else {
+            if ($subItem instanceof ItemInterface) {
+                $subItem = array($subItem);
+                $subItem = array(array(
+                    'label' => $subItem->getLabel(),
+                    'uri' => $subItem->getUri(),
+                    'item' => $subItem,
+                ));
             }
-            $subItem = array_reverse($subItem);
+            if (null === $subItem) {
+                $subItem = array();
+            }
+            if (!is_array($subItem) && !$subItem instanceof \Traversable) {
+                $subItem = array($subItem);
+            }
+
             foreach ($subItem as $key => $value) {
-                if (is_numeric($key)) {
-                    $key = $value;
-                    $value = null;
+                switch (true) {
+                    case $value instanceof ItemInterface:
+                        $value = array(
+                            'label' => $value->getLabel(),
+                            'uri' => $value->getUri(),
+                            'item' => $value,
+                        );
+                        break;
+                    case is_array($value):
+                        // Assume we already have the appropriate array format for the element
+                        break;
+                    case is_integer($key):
+                        $value = array(
+                            'label' => (string) $value,
+                            'uri' => null,
+                            'item' => null,
+                        );
+                        break;
+                    case is_scalar($value):
+                        $value = array(
+                            'label' => (string) $key,
+                            'uri' => (string) $value,
+                            'item' => null,
+                        );
+                        break;
+                    case null === $value:
+                        $value = array(
+                            'label' => (string) $key,
+                            'uri' => null,
+                            'item' => null,
+                        );
+                        break;
+                    default:
+                        throw new \InvalidArgumentException(sprintf('Invalid value supplied for the key "%s". It should be an item, an array or a scalar', $key));
                 }
-                $breadcrumbs[(string) $key] = $value;
+                $breadcrumbs[] = $value;
             }
         }
 
-        do {
-            $label = $obj->getLabel();
-            $breadcrumbs[$label] = $obj->getUri();
-        } while ($obj = $obj->getParent());
+        array_unshift($breadcrumbs, array(
+            'label' => $this->getLabel(),
+            'uri' => $this->getUri(),
+            'item' => $this,
+        ));
 
-        return array_reverse($breadcrumbs, true);
+        if ($this->isRoot()) {
+            return $breadcrumbs;
+        }
+
+        return $this->getParent()->getBreadcrumbsArray($breadcrumbs, true);
     }
 
     /**